# openapi.yaml
openapi: 3.1.0
info:
  title: Lume CRM API
  version: 1.0.0
  description: |
    API specification for Lume CRM - a customer relationship management system for window film sales.
    
    ## Authentication
    
    The API uses cookie-based authentication with JWT tokens:
    - `crm-token` - Short-lived access token (15 min)
    - `crm-refresh-token` - Long-lived refresh token (7 days)
    
    Alternatively, Bearer tokens can be used via the `Authorization` header.
    
    ## Rate Limiting
    
    Login endpoint is rate-limited to 10 requests per 15 minutes per IP.
    
    ## Error Responses
    
    All endpoints return consistent error responses:
    ```json
    {
      "error": "string",
      "code": "string",
      "details": "string",
      "hint": "string"
    }
    ```
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://lume-crm.vercel.app
    description: Production server
security:
  - cookieAuth: []
  - bearerAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: crm-token
      description: JWT access token stored in httpOnly cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token in Authorization header
  schemas:
...
    LoginRequest:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          description: Email or username (will be normalized to email format)
        password:
          type: string
          format: password
          description: User password
        redirectTo:
          type: string
          description: Optional redirect path after successful login
      example:
        login: "vendedor@lume.local"
        password: "senha123"
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        session:
          type: object
          nullable: true
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
      example:
        success: true
        session:
          accessToken: "eyJhbGciOiJIUzI1NiIs..."
          refreshToken: "eyJhbGciOiJIUzI1NiIs..."
    LoginErrorResponse:
      type: object
      properties:
        error:
          type: string
        retryAfter:
          type: integer
          description: Seconds until rate limit resets
      example:
        error: "Muitas tentativas de login. Tente novamente mais tarde."
        retryAfter: 900
...
    MeResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        email:
          type: string
      example:
        authenticated: true
        email: "vendedor@lume.local"
    MeErrorResponse:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Token nao fornecido"
...
    RefreshResponse:
      type: object
      properties:
        success:
          type: boolean
        accessToken:
          type: string
      example:
        success: true
        accessToken: "eyJhbGciOiJIUzI1NiIs..."
    RefreshErrorResponse:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Refresh token invalido ou expirado"

    # Lead schemas
    LeadStatus:
      type: string
      enum: ["Novo", "Em Contato", "Agendado", "Fechado", "Perdido"]
      description: Lead pipeline status
...
    FollowUpPlaybookRule:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/LeadStatus'
        days:
          type: integer
          minimum: 0
        description:
          type: string
    SellerPlaybook:
      type: object
      properties:
        sellerId:
          type: string
        rules:
          type: array
          items:
            $ref: '#/components/schemas/FollowUpPlaybookRule'
    PlaybookResponse:
      type: object
      properties:
        activeSellerId:
          type: string
        activePlaybook:
          $ref: '#/components/schemas/SellerPlaybook'
        sellerIds:
          type: array
          items:
            type: string
    PlaybookUpdateRequest:
      type: object
      properties:
        sellerId:
          type: string
          default: "equipe-lume"
        rules:
          type: array
          items:
            $ref: '#/components/schemas/FollowUpPlaybookRule'

    # Lead Status History schemas
    LeadStatusHistoryEntry:
      type: object
      properties:
        id:
          type: string
        lead_id:
          type: string
        from_status:
          $ref: '#/components/schemas/LeadStatus'
          nullable: true
        to_status:
          $ref: '#/components/schemas/LeadStatus'
        changed_at:
          type: string
          format: date-time
        changed_by:
          type: string
          nullable: true
    LeadStatusHistoryResponse:
      type: array
      items:
        $ref: '#/components/schemas/LeadStatusHistoryEntry'

    # Generic error response
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: string
        hint:
          type: string
  parameters:
...
    playbookSellerIdParam:
      name: sellerId
      in: query
      schema:
        type: string
        default: "equipe-lume"
      description: Seller identifier for playbook
    historyLeadIdParam:
      name: leadId
      in: query
      required: true
      schema:
        type: string
      description: Lead ID to fetch history for
paths:
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: |
        Authenticates user with email/password and sets httpOnly cookies with JWT tokens.
        Supports both JSON and form-encoded requests.
        Rate limited to 10 requests per 15 minutes per IP.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Sets crm-token and crm-refresh-token cookies
            X-RateLimit-Limit:
              schema:
                type: string
            X-RateLimit-Remaining:
              schema:
                type: string
        '303':
          description: Form submission redirect on success
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              schema:
                type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginErrorResponse'
          headers:
            X-RateLimit-Limit:
              schema:
                type: string
            X-RateLimit-Remaining:
              schema:
                type: string
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginErrorResponse'
          headers:
            Retry-After:
              schema:
                type: string
            X-RateLimit-Limit:
              schema:
                type: string
            X-RateLimit-Remaining:
              schema:
                type: string
            X-RateLimit-Reset:
              schema:
                type: string
...
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Returns authenticated user info from JWT token (cookie or Bearer header)
      operationId: getCurrentUser
      responses:
        '200':
          description: User authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeErrorResponse'
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Uses refresh token from cookie to issue new access token pair
      operationId: refreshToken
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Updates crm-token and crm-refresh-token cookies
        '401':
          description: Invalid or missing refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshErrorResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Clears invalid cookies
...
  /api/crm/playbooks:
    get:
      tags:
        - Playbooks
      summary: Get playbooks
      description: |
        Retrieves all seller playbooks and identifies the active one for the requested seller.
        Requires Supabase service role key (admin access).
      operationId: getPlaybooks
      parameters:
        - $ref: '#/components/parameters/playbookSellerIdParam'
      responses:
        '200':
          description: Playbooks retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybookResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Playbooks table not created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error or admin not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Playbooks
      summary: Update playbook
      description: |
        Creates or updates a seller's follow-up playbook rules.
        Requires Supabase service role key (admin access).
      operationId: updatePlaybook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaybookUpdateRequest'
      responses:
        '200':
          description: Playbook updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybookResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Playbooks table not created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error or admin not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/crm/lead-status-history:
    get:
      tags:
        - Lead Status History
      summary: Get lead status history
      description: |
        Retrieves the status change audit trail for a specific lead.
        Requires Supabase service role key (admin access).
      operationId: getLeadStatusHistory
      parameters:
        - $ref: '#/components/parameters/historyLeadIdParam'
      responses:
        '200':
          description: Status history entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadStatusHistoryResponse'
        '400':
          description: Missing leadId parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error or admin not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'